feat : 탈퇴한 회원 테이블 추가 후 기능 구현, 버그픽스

+++ run/src/main/java/com/guide/run/event/service/EventService.java
@@ -17,15 +17,22 @@ import com.guide.run.event.entity.type.EventRecruitStatus;
 import com.guide.run.event.entity.type.EventStatus;
 import com.guide.run.global.converter.TimeFormatter;
 import com.guide.run.global.exception.event.authorize.NotEventOrganizerException;
+import com.guide.run.global.exception.event.dto.NotValidEventRecruitException;
+import com.guide.run.global.exception.event.dto.NotValidEventStartException;
+import com.guide.run.global.exception.event.logic.NotDeleteEventException;
 import com.guide.run.global.exception.event.resource.NotExistEventException;
 import com.guide.run.global.exception.user.resource.NotExistUserException;
 import com.guide.run.partner.entity.matching.Matching;
 import com.guide.run.partner.entity.matching.repository.MatchingRepository;
+import com.guide.run.partner.entity.partner.Partner;
+import com.guide.run.partner.entity.partner.repository.PartnerRepository;
+import com.guide.run.temp.member.repository.AttendanceRepository;
 import com.guide.run.user.entity.type.Role;
 import com.guide.run.user.entity.type.UserType;
 import com.guide.run.user.entity.user.User;
 import com.guide.run.user.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.apache.poi.sl.image.ImageHeaderBitmap;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -46,6 +53,8 @@ public class EventService {
     private final MatchingRepository matchingRepository;
     private final TimeFormatter timeFormatter;
     private final EventLikeRepository eventLikeRepository;
+
+    private final AttendanceRepository attendanceRepository;
     @Transactional
     public EventCreatedResponse eventCreate(EventCreateRequest request, String privateId){
         User user = userRepository.findUserByPrivateId(privateId).
@@ -57,7 +66,7 @@ public class EventService {
                 .recruitEndDate(request.getRecruitEndDate())
                 .name(request.getTitle())
                 .recruitStatus(EventRecruitStatus.RECRUIT_UPCOMING)
-                .isApprove(false)
+                .isApprove(true)
                 .type(request.getEventType())
                 .startTime(timeFormatter.getDateTime(request.getDate(), request.getStartTime()))
                 .endTime(timeFormatter.getDateTime(request.getDate(), request.getEndTime()))
@@ -72,6 +81,10 @@ public class EventService {
                         .privateIds(new ArrayList<>())
                         .build()
         );
+
+        validEventTime(createdEvent);
+
+
         return EventCreatedResponse.builder()
                 .eventId(createdEvent.getId())
                 .isApprove(createdEvent.isApprove())
@@ -100,23 +113,27 @@ public class EventService {
                     .maxNumG(request.getMinNumG())
                     .place(request.getPlace())
                     .content(request.getContent()).build());
+
+            validEventTime(updatedEvent);
+
             return EventUpdatedResponse.builder()
                     .eventId(updatedEvent.getId())
                     .isApprove(updatedEvent.isApprove())
                     .build();
+
         }
         throw new NotEventOrganizerException();
     }
 
     //이벤트 모집 마감
     @Transactional
-    public void eventClose(String userId, Long eventId){
-        User user = userRepository.findUserByPrivateId(userId).
+    public void eventClose(String privateId, Long eventId){
+        User user = userRepository.findUserByPrivateId(privateId).
                 orElseThrow(NotExistUserException::new);
 
         Event event = eventRepository.findById(eventId).orElseThrow(NotExistEventException::new);
 
-        if(!event.getOrganizer().equals(userId)){
+        if(!event.getOrganizer().equals(privateId)){
             throw new NotEventOrganizerException();
         }
         //지금 시간 긁어서 이벤트 마감 시간으로 변경 + 이벤트 상태 변경.
@@ -130,12 +147,17 @@ public class EventService {
         Event event = eventRepository.findById(eventId).orElseThrow(NotExistEventException::new);
         User user = userRepository.findUserByPrivateId(userId).orElseThrow(NotExistUserException::new);
 
+        if(event.getStatus().equals(EventStatus.EVENT_END)){
+            throw new NotDeleteEventException();
+        }
         if(event.getOrganizer().equals(userId) || user.getRole().equals(Role.ROLE_ADMIN)){
             //todo : 매칭, 출석, 신청서, 파트너, 이벤트 좋아요, 댓글 전부 삭제해야 함.
-            eventFormRepository.deleteAllByEventId(eventId);
 
-            eventRepository.deleteById(eventId);
+            eventFormRepository.deleteAllByEventId(eventId);
+            matchingRepository.deleteAllByEventId(eventId);
+            attendanceRepository.deleteAllByEventId(eventId);
             eventLikeRepository.deleteById(eventId);
+            eventRepository.deleteById(eventId);
         }
         else
             throw new NotEventOrganizerException();
@@ -152,10 +174,13 @@ public class EventService {
         );
 
         boolean apply = false;
+
+        boolean isMatching = false;
         //신청 여부
         EventForm eventForm = eventFormRepository.findByEventIdAndPrivateId(eventId, privateId);
         if(eventForm!=null){
             apply = true;
+            isMatching = eventForm.isMatching();
         }
 
 
@@ -174,7 +199,7 @@ public class EventService {
                 .updatedAt(LocalDate.from(event.getUpdatedAt()))
                 .isApply(apply)
                 //todo : 2차에서 추가된 부분
-                .hasPartner(eventForm.isMatching()) //파트너 존재 여부
+                .hasPartner(isMatching) //파트너 존재 여부
                 .partnerName(null) //파트너 이름
                 .partnerRecord(null) //파트너 러닝등급
                 .partnerType(null) //파트너 장애여부
@@ -183,7 +208,7 @@ public class EventService {
         //매칭 여부로 파트너 정보 추가
         Matching matching;
         String partnerId;
-        if(eventForm.isMatching()){
+        if(isMatching){
             if(user.getType().equals(UserType.GUIDE)){
                 matching = matchingRepository.findByEventIdAndGuideId(eventId, privateId);
                 partnerId = matching.getViId();
@@ -211,4 +236,17 @@ public class EventService {
     public DetailEvent getDetailEvent(Long eventId, String userId) {
         return new DetailEvent();
     }
+
+    public void validEventTime(Event event){
+        if(event.getRecruitStartDate().isAfter(event.getRecruitEndDate())
+                || event.getRecruitStartDate().isEqual(event.getRecruitEndDate())){
+            throw new NotValidEventRecruitException();
+        }
+
+        if(event.getStartTime().isAfter(event.getEndTime())
+                ||event.getStartTime().isEqual(event.getEndTime())){
+            throw new NotValidEventStartException();
+        }
+    }
+
 }
